Fix up mkinitrd scripts to detect bnx2fc driver
The original mkinitrd scripts only checked for the intel-style
sysfs layout, where the SCSI host was located under the nic.
Broadcom is placing the SCSI host under the PCI device, so
the scripts need to be updated to copy with that.
And while we're at it, I've removed the dependency on the
network scripts, so all NIC drivers are now included, too.

References: bnc#725579

Signed-off-by: Hannes Reinecke <hare@suse.de>