Remove PID file usage
Using a PID file is not recommended for daemons serving the root
filesystem, as the required /var/run directories might not be
mounted initially.

And as we're communicating via netlink sockets with the daemon
anyway we can as well add a new command for fcoeadm for printing
out the PID of the daemon.

Signed-off-by: Hannes Reinecke <hare@suse.de>