fipvlan: fix broken fcoe start using fipvlan
FCoE start is broken due to the link being never selected and that
is in turn due to bad vlan interface name compare against real
interface name passed for fip vlan discovery.

This name compare is unnecessary and breaks fipvlan start option,
so this patch removes it to have fcoe start working again.

I considered reducing compare to just the base real interface name
but that won't work in the case of existing vlan interface name is
already customized, so instead just don't check for the name.

I verified this fixes the start using newly created vlan as well
as existing vlan, which could be down or already running.

Signed-off-by: Vasu Dev <vasu.dev@intel.com>